Workers can build multiple queued walls
Right before the FSM redesign attempt.

Worker moves to the first wall segment and continues with the next.

There's a problem with CommandQueue in the SFM component - the second worker with their own wall moves to the walls of the first worker. He get's a copy of build tasks of the previous worker